backend/requirements.txt
# Django Core
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Environment variables
python-decouple>=3.8,<4.0

# Development tools
django-debug-toolbar>=4.2,<5.0

# Utilities
pytz>=2023.3
